feat: break GridOverlay into it's own file since it is "use client"
